A3 Wasteland Stratis
Wasteland.Stratis

Original version author CRE4MPIE

Edits wernerz, RV

A3Wasteland Stratis (A3Wasteland ReadMe)

A3Wasteland is a survival-like combat sandbox, which incorporates
elements of team deathmatch and sector control in one big fun-to-play
package. Similar to DayZ, the objectives are to scavenge weapons and
vehicles, team up with fellow players, and engage in combat to obtain
control of the landscape. Best experienced with 50 or more players.

The source code is based on GoT Wasteland, which is based on 404
Wasteland, which is based on Sa-Matra's Wasteland, which is finally
based on Tonic's Wasteland.

The mission has come a long way since 404 Wasteland. A large portion of
the code a has been rewritten, and it contains many new features like
territory capture, player saving, vehicle saving, vehicle stores, new
secondary missions, and more.

For questions and support, please visit our forums.

Team Wasteland developers

AgentRev
[GoT] JoSchaap
MercyfulFate
[KoS] His_Shadow
[KoS] Bewilderbeest
Torndeco
[404] Del1te
A list of other contributors can be found in the "Credits" section from
the in-game map menu, or near the bottom of briefing.sqf

This project is licensed under the GNU Affero GPL v3. Copyright © 2013 -
2015 A3Wasteland.com